move new governor functions to gov_lib
 * setGovRobot and waitGov
